refactor(sputnik): group js impls db types
