+++ src/sputnik/src/hooks/js/impls/utils.rs
@@ -1,5 +1,4 @@
-use crate::hooks::js::impls::JsBigInt;
-use crate::hooks::js::types::hooks::JsDoc;
+use crate::hooks::js::impls::shared::JsBigInt;
 use rquickjs::{BigInt, Ctx, Error as JsError, IntoJs, Result as JsResult, Value};
 
 pub fn into_optional_bigint_js<'js>(ctx: &Ctx<'js>, value: Option<u64>) -> JsResult<Value<'js>> {
@@ -25,15 +24,5 @@ pub fn from_optional_bigint_js(value: Option<BigInt>) -> JsResult<Option<u64>> {
             }
         }
         None => Ok(None),
-    }
-}
-
-pub fn into_optional_jsdoc_js<'js>(
-    ctx: &Ctx<'js>,
-    value: Option<JsDoc<'js>>,
-) -> JsResult<Value<'js>> {
-    match value {
-        Some(value) => value.into_js(ctx),
-        None => Ok(Value::new_undefined(ctx.clone())),
     }
 }
