+++ src/sputnik/src/hooks/js/mod.rs
@@ -3,4 +3,3 @@ pub mod loaders;
 pub mod runtime;
 pub mod sdk;
 mod types;
-mod utils;
